+++ Modules/Intent.Modules.Modelers.Services.DomainInteractions/resources/crud-api.js
@@ -543,25 +543,37 @@ async function openCrudCreationDialog(preselectedClassId, defaultDiagramId) {
                     fieldType: "tree-view",
                     label: "Domain Operations",
                     hint: "Generate operations from selected domain entity operations",
+                    isRequired: false,
                     treeViewOptions: {
-                        rootId: foundEntity.id,
-                        submitFormTriggers: ["double-click", "enter"],
+                        rootNode: {
+                            specializationId: "Class",
+                            id: "root",
+                            label: "Loading...",
+                            isExpanded: true,
+                            children: [],
+                        },
                         isMultiSelect: true,
+                        submitFormTriggers: ["double-click", "enter"],
                         selectableTypes: [
                             {
                                 specializationId: "Class",
                                 autoExpand: true,
                                 autoSelectChildren: false,
-                                isSelectable: (x) => false
+                                isSelectable: false
                             },
                             {
                                 specializationId: "Operation",
-                                isSelectable: (x) => true
+                                isSelectable: true
                             }
                         ]
                     }
                 }
-            ]
+            ],
+            onInitialize: async (formApi) => {
+                // This somehow loads the tree correctly as opposed to populating in the structure above directly.
+                const tree = formApi.getField("tree");
+                tree.treeViewOptions.rootId = foundEntity.id;
+            }
         });
         result.selectedDomainOperationIds = (_b = (_a = dialogResult.tree) === null || _a === void 0 ? void 0 : _a.filter((x) => x != "0")) !== null && _b !== void 0 ? _b : [];
     }