+++ Modules/Intent.Modules.Modelers.Services.DomainInteractions/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 2.3.7
 
 - Improvement: It is now possible to use the result of an invocation as a data source on `Create` and `Update` mapping screens.
+- Fixed: Create CRUD Operations for Services and CQRS didn't generate elements for Domain Entity Operations.
 
 ### Version 2.3.6